section08/115_migrate_to_authlib/end/requirements.txt
psycopg2-binary
Flask
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-OAuthlib
Flask-marshmallow
python-dotenv
marshmallow-sqlalchemy
marshmallow
authlib